Toggle to store JSON files scanItems.json and scanData.json (#424)
* Prettier linting changes

* Add a cli option to generate json file for scanItems and scanData

* Update README that introduces -g flag to CLI

* Bump package version

---------

+++ src/mergeAxeResults.ts
@@ -457,14 +457,14 @@ function writeLargeJsonToFile(obj, filePath) {
   });
 }
 
-const base64Encode = async (data, num) => {
+const base64Encode = async (data, num, storagePath, generateJsonFiles) => {
   try {
     const tempFilename =
       num === 1
-        ? `scanItems_${uuidv4()}.json`
+        ? path.join(storagePath, 'scanItems.json')
         : num === 2
-          ? `scanData_${uuidv4()}.json`
-          : `${uuidv4()}.json`;
+          ? path.join(storagePath, 'scanData.json')
+          : path.join(storagePath, `${uuidv4()}.json`);
     const tempFilePath = path.join(process.cwd(), tempFilename);
 
     await writeLargeJsonToFile(data, tempFilePath);
@@ -492,9 +492,11 @@ const base64Encode = async (data, num) => {
 
       return outputFilePath;
     } finally {
+      if (!generateJsonFiles) {
       await fs.promises
         .unlink(tempFilePath)
         .catch(err => console.error('Temp file delete error:', err));
+      }
     }
   } catch (error) {
     console.error('Error encoding data to Base64:', error);
@@ -520,10 +522,10 @@ const streamEncodedDataToFile = async (inputFilePath, writeStream, appendComma) 
   }
 };
 
-const writeBase64 = async (allIssues, storagePath) => {
+const writeBase64 = async (allIssues, storagePath, generateJsonFiles) => {
   const { items, ...rest } = allIssues;
-  const encodedScanItemsPath = await base64Encode(items, 1);
-  const encodedScanDataPath = await base64Encode(rest, 2);
+  const encodedScanItemsPath = await base64Encode(items, 1, storagePath, generateJsonFiles);
+  const encodedScanDataPath = await base64Encode(rest, 2, storagePath, generateJsonFiles);
 
   const filePath = path.join(storagePath, 'scanDetails.csv');
   const directoryPath = path.dirname(filePath);
@@ -766,6 +768,7 @@ const generateArtifacts = async (
   cypressScanAboutMetadata,
   scanDetails,
   zip = undefined, // optional
+  generateJsonFiles = false,
 ) => {
   const intermediateDatasetsPath = `${randomToken}/datasets/${randomToken}`;
   const phAppVersion = getVersion();
@@ -908,7 +911,7 @@ const generateArtifacts = async (
   }
 
   await writeCsv(allIssues, storagePath);
-  await writeBase64(allIssues, storagePath);
+  await writeBase64(allIssues, storagePath, generateJsonFiles);
   await writeSummaryHTML(allIssues, storagePath);
   await writeHTML(allIssues, storagePath);
   await retryFunction(() => writeSummaryPdf(storagePath, pagesScanned.length), 1);
